+++ extension/customize/components/controls/Slider.js
